+++ packages/cli/routines/check_schema.js
@@ -33,7 +33,7 @@ export async function main( { filename }, environment, dry, verbose, printJson )
       return;
     }
 
-    store._schema_fields.forEach(field => {
+    store.schemaFields.forEach(field => {
       if ( field.validation_bolt_type || field.validate_bolt_type ) {
         console.error('['+storeName+' -> '+field.model+']', chalk.red("validation_bolt_type/validate_bolt_type should be type"))
       }